feat: add PageEnterMotion component for enhanced entrance animations
